--- apps/nextjs-website/react-components/src/index.ts
@@ -1,4 +1,0 @@
-import Hero from './components/Hero/Hero';
-import Editorial from './components/Editorial/Editorial';
-
-export { Hero, Editorial };
