+++ apps/nextjs-website/react-components/types/index.ts
@@ -1,3 +1,4 @@
 import { HeroProps } from './Hero/Hero.types';
+import { EditorialProps } from './Editorial/Editorial.types';
 
-export type { HeroProps };
+export type { HeroProps, EditorialProps };
