+++ apps/nextjs-website/src/components/Editorial.tsx
@@ -1,8 +1,8 @@
 'use client';
 import Image from 'next/image';
 import MarkdownRenderer from './MarkdownRenderer';
-import { Editorial as EditorialEC } from '@react-components';
-import { EditorialProps } from '@react-components-props';
+import { Editorial as EditorialEC } from '@react-components/components';
+import { EditorialProps } from '@react-components/types';
 import { EditorialSection } from '@/lib/fetch/types/PageSection';
 import Icon from '@/components/Icon';
